1. Added Admin order page 2. Order filter 3. Order delete, change status functionality 4. Added NotFound functionality when user is admin or simple user 5. Deleted price and quality ui's

+++ src/views/admin/AdminProduct.vue
@@ -94,7 +94,7 @@ export default {
       confirm.value = true
     }
     watch(flag, async (response) => {
-      if (response) {
+      if (response && deleteFlag.value) {
         await store.dispatch('products/deleteProduct', product.value.id)
         await router.push('/admin/products')
       }
